JE2-187. Created Users Page, fixed column options display in english
